+++ src/main/java/github/grace000/erversay/RouteHandlers/Options.java
@@ -1,20 +1,27 @@
-package github.grace000.erversay.Handlers;
+package github.grace000.erversay.RouteHandlers;
 
 import github.grace000.erversay.Request;
 import github.grace000.erversay.ResponseBuilder;
 
+import static github.grace000.erversay.Constants.Headers.OPTIONS_HEADER;
 import static github.grace000.erversay.Constants.StatusCodes.NOT_FOUND_STATUS;
 
-public class Post implements Handler {
+public class Options implements RouteHandler{
+    public enum AcceptedMethods {
+        GET, HEAD, OPTIONS
+    }
+
     public boolean isMethodAllowed(String method) {
-        return method.equals("POST");
+        for (AcceptedMethods accepted: AcceptedMethods.values()) {
+            return (accepted.name().equals(method));
+        }
+        return false;
     }
 
     public String handle(Request request) {
         if (isMethodAllowed(request.method)) {
             return new ResponseBuilder()
-                    .withBody(request.body)
-                    .withContentLength(request.body.length(), request.body)
+                    .withHeaders(OPTIONS_HEADER)
                     .build();
         } else return new ResponseBuilder()
                 .withStatus(NOT_FOUND_STATUS)
