+++ src/main/java/github/grace000/erversay/RouteHandlers/Post.java
@@ -1,29 +1,20 @@
-package github.grace000.erversay.Handlers;
+package github.grace000.erversay.RouteHandlers;
 
 import github.grace000.erversay.Request;
 import github.grace000.erversay.ResponseBuilder;
 
-import static github.grace000.erversay.Constants.Headers.OPTIONS_TWO_HEADER;
 import static github.grace000.erversay.Constants.StatusCodes.NOT_FOUND_STATUS;
 
-public class OptionsTwo implements Handler {
-    public enum AcceptedMethods {
-        GET, HEAD, OPTIONS, PUT, POST
-    }
-
+public class Post implements RouteHandler {
     public boolean isMethodAllowed(String method) {
-        for (AcceptedMethods accepted: AcceptedMethods.values()) {
-            if (accepted.name().equals(method)){
-                return true;
-            }
-        }
-        return false;
+        return method.equals("POST");
     }
 
     public String handle(Request request) {
         if (isMethodAllowed(request.method)) {
             return new ResponseBuilder()
-                    .withHeaders(OPTIONS_TWO_HEADER)
+                    .withBody(request.body)
+                    .withContentLength(request.body.length(), request.body)
                     .build();
         } else return new ResponseBuilder()
                 .withStatus(NOT_FOUND_STATUS)
